+++ includes/Helper.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace PagarmeSplitPayment;
+
+class Helper
+{
+    public static function priceInCents(float $price)
+    {
+        return $price * 100;
+    }
+}