+++ includes/Pagarme/SplitRules.php
@@ -2,9 +2,12 @@
 
 namespace PagarmeSplitPayment\Pagarme;
 
+use PagarmeSplitPayment\Entities\Partner;
+use PagarmeSplitPayment\Helper;
+
 class SplitRules {
     public function split( $data, $order ) {
-        $partners = $this->partnersPercentageOverOrder($order);
+        $partners = $this->partnersAmountOverOrder($order);
         $mainRecipientData = carbon_get_theme_option('psp_partner');
 
         if (
@@ -15,7 +18,8 @@ class SplitRules {
             return $data;
         }
 
-        $partnersPercentage = 0;
+        $partnersAmount = 0;
+
         foreach($partners as $id => $partner) {
             $partnerData = carbon_get_user_meta($id, 'psp_partner')[0];
             
@@ -23,70 +27,68 @@ class SplitRules {
                 continue;
             }
 
-            // Count the percentage splited to partners
-            $partnersPercentage += $partner['percentage'];
+            $partnersAmount += Helper::priceInCents($partner['value']);
 
             $data['split_rules'][] = [
                 'recipient_id' => $partnerData['psp_recipient_id'],
-                'percentage' => $partner['percentage'],
+                'amount' => Helper::priceInCents($partner['value']),
                 'liable' => true,
                 'charge_processing_fee' => true,
             ];
         }
 
         // If there is no percentage to split return original data
-        if (!$partnersPercentage) {
+        if (!$partnersAmount) {
             return $data;
         }
 
         $data['split_rules'][] = [
             'recipient_id' => $mainRecipientData[0]['psp_recipient_id'],
-            'percentage' => 100 - $partnersPercentage,
+            'amount' => Helper::priceInCents($order->get_total()) - $partnersAmount,
             'liable' => true,
             'charge_processing_fee' => true,
         ];
 
         $this->log($order);
-    
+
         return $data;
     }
 
     /**
-	 * Calculate the percentage that each partner should receive over the order 
-     * based on the values he should receive over each product
-	 *
+	 * Calculate the amount that each partner should receive over the order
+     *
 	 * @param mixed $order WooCommerce Order object.
 	 * @return array
 	 */
     // 
-    private function partnersPercentageOverOrder($order)
+    private function partnersAmountOverOrder(\WC_Order $order)
     {
         $items = $order->get_items();
         $partners = [];
-        foreach ( $items as $item ) {
-            $productPartners = carbon_get_post_meta(
-                $item->get_product_id(), 
-                'psp_partners'
-            );
 
-            // Sum the total amount to be given to each partner on the order
-            foreach ($productPartners as $partner) {
-                $partners[$partner['psp_partner_user'][0]['id']]['value'] += (
-                    $item->get_data()['total'] * ($partner['psp_percentage']/100)
-                );
+        foreach ($items as $item) {
+            foreach ($this->getPartnersFromProduct($item->get_product_id()) as $partner) {
+                $userId = (int) $partner['psp_partner'][0]['id'];
+                $partner = new Partner($userId);
+                $partners[$userId]['value'] += $partner->calculateComission($item)->getComission();
             }
         }
 
-        $productsTotal = $order->get_subtotal();
-
-        foreach($partners as $id => $partner) {
-            // Round the value because pagarme doesnt allow float
-            $partners[$id]['percentage'] = round(
-                ($partner['value'] * 100) / $productsTotal
-            );
-        }
-
         return $partners;
+    }
+
+    private function getPartnersFromProduct(int $productId): array {
+        $partners = [
+            'percentage' => carbon_get_post_meta($productId, 'psp_percentage_partners'),
+            'fixed_amount' => [[
+                'psp_partner' => carbon_get_post_meta($productId, 'psp_fixed_partner'), 
+                'psp_comission_value' => carbon_get_post_meta($productId, 'psp_comission_value')
+            ]]
+        ];
+
+        $comissionType = carbon_get_post_meta($productId, 'psp_comission_type');
+
+        return $partners[$comissionType];
     }
 
     private function log($order)