+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.1.0] - 2022-04-27
+### Added
+- Possibility to define fixed amount for a partner in the payment split
+
+### Fixed
+- Splited amounts when using a discount coupon in the order
+
 ## [1.0.2] - 2021-12-29
 ### Changed
 - Restrict access to admin area only for roles that explicit this behavior